fix: remove package dependency on dev_scripts
deribit_wrapper.account_management imported create_multilevel_df from
dev_scripts, making the published package depend on a folder of debug
scripts that also gets shipped in the wheel via find_packages().

Move flatten_dict and create_multilevel_df into
deribit_wrapper.utilities where they belong; dev_scripts.utilities_dev
now re-exports them for backward compatibility.
